Fix memory safety and lock-discipline defects in parallel hash aggregation
Eleven fixes, all found by review of the preceding commits, none of them about
speed.  Grouped by what they protect.

Shared-memory safety:

- shared_agg_alloc_entry() sized an oversized entry's dedicated chunk from the
  bare header while carving the first entry at CACHELINEALIGN(header), so any
  entry above SHARED_AGG_CHUNK_SIZE - PG_CACHE_LINE_SIZE ran past the end of
  its dsa_allocate().  Silent corruption of a neighbouring DSA object, in a
  process other than the one that would read it.  Size the chunk for where the
  carve actually starts, and Assert the carve fits -- the assertion matters
  more than the arithmetic, because nothing else would notice.

- Clamp numGroups in double before casting it to uint64 when sizing the bucket
  array.  It is only clamped to MAXIMUM_ROWCOUNT (1e100), and the cast is
  undefined behaviour above PG_UINT64_MAX; pg_nextpower2_64() would then assert
  or shift by 64.

- mem_limit is clamped to PG_UINT64_MAX to mean "no limit", and the spill test
  compared it as int64 -- where that value is -1, putting the table into spill
  mode on its first insert.  Compare in uint64, with a sign test because the
  running total can be transiently negative between flushes.

Nothing that can block, take a heavyweight lock or do I/O under a stripe lock:

- Grouping keys were stored in shared memory with their TOAST pointers intact:
  prepare_hash_slot() copies datums verbatim and heap_fill_tuple() preserves an
  external pointer, so the locked chain walk in shared_agg_insert() could reach
  detoast_attr() -- table_open(toastrel, AccessShareLock), an index scan, buffer
  pins and possibly physical I/O -- with interrupts held off and every
  participant hashing to that stripe queued behind it.  This is the hazard
  SharedAggTransInfo.arg_detoast already avoids for the aggregate's arguments;
  the key side had no equivalent.  Add shared_agg_detoast_key(), called from
  both the build and the refill loop.

- Warm the collation cache for the grouping keys' and the aggregates' input
  collations in shared_agg_init_support().  texteq() and varstr_cmp() call
  pg_newlocale_from_collation() on every call, and for a non-default collation
  the first call does a syscache lookup -- and for ICU opens the collator.

- Refuse array transition types.  min(anyarray) and max(anyarray) resolve to a
  concrete array of a by-value element and passed every other test, while
  array_smaller() reaches array_cmp(), which calls the element type's
  comparison procedure -- PL/pgSQL, for a user-defined by-value base type.  The
  gate can only see the type, and "a flat int8[] is only read by memcpy" is
  true of int4_avg_accum, not of transition functions in general.  This costs
  avg(int2), avg(int4) and the float8 statistics family; readmitting them wants
  a test on the transition function rather than the type.

Wrong answers:

- The "parallel-aware Agg requires the hashed strategy" cross-check sat inside
  the use_hashing block, which AGG_SORTED and AGG_PLAIN never enter -- so it
  caught only AGG_MIXED and missed precisely the two strategies that would fall
  through to the private path and run AGGSPLIT_SIMPLE over a partial input.
  Move it to the top of ExecInitAgg().

Resource release:

- Release the shared table on the completion path, not only on rescan.  The
  last participant to detach from the scan barrier frees it, which is the same
  point PHJ frees at (PHJ_BUILD_FREE) and for the same reason: a finished table
  otherwise occupies es_query_dsa until the Gather shuts down.  Report the
  bucket count from a separate field so EXPLAIN still has it afterwards.

- Delete the previous run's spill files in ExecAggReInitializeDSM().  The
  comment claimed this would require tracking their names; it does not --
  SharedFileSetDeleteAll() takes none, and ExecHashJoinReInitializeDSM() calls
  it in exactly the analogous place.

Correctness of the bucket index, and tidying:

- Compute the bucket index by rotating the hash rather than masking it.  Every
  row in a spill batch agrees in its top cur_used_bits hash bits, and the
  bucket array never shrinks, so a plain low-bit mask makes the two ranges
  overlap once cur_used_bits + log2(nbuckets) passes 32: for a 2^23-bucket
  table, half the array unreachable at depth two and 63/64 of it at depth
  three, with the chains correspondingly longer.

- Refuse a child grouped rel outright in parallel_shared_hashagg_possible().
  Under partitionwise aggregation the statistics test there was reading the
  partitioned parent's distribution to judge a child, which is the same
  evidence the cost model twenty lines further on declines to trust.  This also
  reflects the honest state of the combination: the resource accounting for N
  shared tables under one Parallel Append is unanalysed.

- enable_parallel_hash_agg's C initialiser now matches its boot_val;
  AggState.shared_nbatches is uint64 like its three siblings; SharedAggBatch is
  in typedefs.list; and shared_agg_reset_table() has its own header comment
  again rather than a stranded one that had become false.

Before posting, these should be folded into the commits that introduced them
rather than left as a fixup on top.

+++ src/include/nodes/execnodes.h
@@ -2520,7 +2520,7 @@ typedef struct AggState
 	 * rather than in SharedAggState because explain.c reads them.
 	 */
 	uint64		shared_nspilled;
-	int			shared_nbatches;
+	uint64		shared_nbatches;
 	uint64		shared_nbuckets;
 	uint64		shared_mem_used;
 } AggState;
